fixed timing cost NaN bug in placer
removed hv * 1.25 Factor in Router: ResourceWithCost

switched timing analysis for placer from int to long (will probably be switched back again, no need to switch in the first place)

+++ task_1/src/designAnalyzer/structures/Net.java
@@ -357,7 +357,7 @@ public class Net {
 	 * updates tR in all simplePaths the sinks of which are part of this net
 	 * @param newTR new value for tR
 	 */
-	public void updateTRAllSinks(int newTR) {
+	public void updateTRAllSinks(long newTR) {
 		for(SimplePath p : sinkingPaths) {
 			p.updateTR(newTR);
 		}
